@@ -13,10 +13,10 @@
  * permissions and limitations under the License.
  */
 
-import crt_native = require('./binding');
+import crt_native from './binding';
 import { isNumber } from 'util';
 
-/** 
+/**
  * Represents an error encountered in native code. Can also be used to convert an error code(Number) into
  * a human-readable string.
  */
