@@ -15,14 +15,14 @@
 
 import * as path from 'path';
 
-let binding: any;
+let binding;
 try { /* when in the lib folder, it's in the dist directory */
-    const binding_path: string = path.resolve(__dirname, '..', '..', 'dist', 'native', 'aws-crt-nodejs');
+    const binding_path = path.resolve(__dirname, '..', '..', 'dist', 'native', 'aws-crt-nodejs');
     binding = require(binding_path);
 }
 catch (err) { /* When in the dist/lib folder, just leave lib */
-    const binding_path: string = path.resolve(__dirname, '..', '..', 'native', 'aws-crt-nodejs');
+    const binding_path = path.resolve(__dirname, '..', '..', 'native', 'aws-crt-nodejs');
     binding = require(binding_path);
 }
 
-export = binding;
+export default binding;
